backend/requirements.txt
flask
flask-cors
transformers
torch
opencv-python
numpy
pillow
requests
# Optional for OCR
paddlepaddle
paddleocr